% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_average_prediction_metrics.R
\name{NMRmetab_average_prediction_metrics}
\alias{NMRmetab_average_prediction_metrics}
\title{compute PLS-DA prediction metrics, VIPs}
\usage{
NMRmetab_average_prediction_metrics(
  dat,
  groupID = "disease_status",
  pred_factor = NULL,
  index_col = 9,
  components_model = 5,
  iterations = 5,
  run_CV = F,
  CV_validation = "loo",
  CV_folds = 5,
  CV_repeats = 100
)
}
\arguments{
\item{dat}{a matrix or a data.frame. rows are samples and columns are grouping and metabolites or bins}

\item{groupID}{a vercors with the column name contining the grouping of the data. This parameter is to be use to mae sure that the 2 new data.frames created have all groups mentioned in the column we identifiy}

\item{pred_factor}{string. name of the group you want ptredction values. if set to null then only fist factor alphabetically ordered is used.}

\item{index_col}{name of the column containing the groups}

\item{components_model}{number of components in the model}

\item{iterations}{number of iterations. ie number of times you are running the random splits -> plsda -> test prediction}

\item{run_CV}{boolean. default is false. intensive computer work. performs the CV on each train split for the maximum number of comomponent.}

\item{CV_validation}{string. either 'loo' default of 'Mfold'}

\item{CV_folds}{integer. number of folds. only used for Mfold validation}

\item{CV_repeats}{integer. number of repeats only used for Mfold validation}
}
\description{
alot of stuff
}
