% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRmetab_normandscale.R
\name{NMRMetab_norm_scale}
\alias{NMRMetab_norm_scale}
\title{Normalisation and Scaling}
\usage{
NMRMetab_norm_scale(
  data,
  index_col = 3,
  normalisation = "None",
  bin = NA,
  scaling = "None",
  writeToFile = F
)
}
\arguments{
\item{data}{a data.frame. Column as variable and rows as sample}

\item{index_col}{integer value. The column number with first bin/matabolite}

\item{normalisation}{string. one of 'None'#default, 'PQN', 'TotArea', 'Bin'}

\item{bin}{a string. name of the bin/metabolite to normalise against}

\item{scaling}{string. one of 'None'#default, 'Auto', 'Pareto', 'Range' or 'Mean'}

\item{writeToFile}{boolean. write resulting data.frame to csv file. F #}
}
\description{
Normalisation and Scaling
}
\author{
Eva Caamano Gutierrez

Arturas Grauslys
}
