% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_average_prediction_metrics.R
\name{NMRMetab_PLS_DA_plot}
\alias{NMRMetab_PLS_DA_plot}
\title{compute PLS-DA scores and plot plsda score}
\usage{
NMRMetab_PLS_DA_plot(
  data,
  groupID = "group",
  index_col = 3,
  components = c(1, 2),
  scale = F,
  elipses = F,
  size_point = 3
)
}
\arguments{
\item{data}{a matrix or a data.frame. rows are samples and columns are grouping and metabolites or bins}

\item{groupID}{a vercors with the column name contining the grouping of the data. This parameter is to be use to mae sure that the 2 new data.frames created have all groups mentioned in the column we identifiy}

\item{index_col}{name of the column containing the groups}

\item{components}{a vector with 2 integer values. default set to 1 and 2}

\item{scale}{does it need scaling}

\item{elipses}{boolean. add elipsess around the groups? defailt is False}

\item{size_point}{double. control the size of the plotted points}
}
\description{
PLSDA score plot
}
